adding getting available time slots. unit tests and integration tests are not working

+++ src/main/java/com/bdd/slots/TimeSlotRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 import java.util.Set;
 
 public interface TimeSlotRepository extends JpaRepository<TimeSlot, Long> {
-    Set<TimeSlot> findAllByActiveAppointmentsIsLessThan(Long stylistsCount);
+    Set<TimeSlot> findAllByActiveAppointmentsIsLessThan(Integer stylistsCount);
     Optional<TimeSlot> findByStartDateAndActiveAppointmentsLessThan(Date startDate, Integer stylistCount);
 
 }
